enoe refactor, geo and etl
